+++ moor_ffi/lib/database.dart
@@ -1,11 +1,9 @@
-/// Exports the low-level [Database] and [IsolateDb] classes to run operations
-/// on a sqflite database.
+/// Exports the low-level [Database] class to run operations on a sqlite
+/// database via `dart:ffi`.
 library database;
 
 import 'package:moor_ffi/src/bindings/types.dart';
-import 'src/impl/isolate/isolate_db.dart';
 
-export 'src/api/database.dart';
 export 'src/api/result.dart';
-export 'src/impl/database.dart' show SqliteException, Database;
-export 'src/impl/isolate/isolate_db.dart';
+export 'src/impl/database.dart'
+    show SqliteException, Database, PreparedStatement;