+++ moor_ffi/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Remove the `background` flag from the moor apis provided by this package. Use the moor isolate api
   instead.
+- Remove builtin support for background execution from the low-level `Database` api
 - Support Dart 2.6, drop support for older versions
 
 ## 0.0.1